+++ tests/compile_fail.rs
@@ -2,5 +2,5 @@
 #[test]
 fn mustnt_compile() {
     let t = trybuild::TestCases::new();
-    t.compile_fail("tests/compile_fail/*.rs");
+    t.compile_fail("tests/compile_fail/**/*.rs");
 }