+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added "nightly-dropck-eyepatch" feature to allow box and vectors to store types that don't outlive it.
 - Improve performance of `extend_from_slice_clone`. ([#69])
 - Improve documentation.